+++ R/Celltype_Verification.R
@@ -13,12 +13,12 @@
 #' @param gene_number Integer specifying number of top genes to select per cell type.
 #' @param annotation_col Character string specifying the column in meta.data to use for grouping.
 #' @param colour_low Color for lowest expression level. (default = "white")
-#' @param colour_high Color for highest expression level. (default = "black")
+#' @param colour_high Color for highest expression level. (default = "navy")
 #'
 #' @return A ggplot object showing expression of top variable genes.
 #'
 #' @export
-#' @family Automated_Annotation_Workflow
+#' @family Section_3_Automated_Annotation_Workflow
 #'
 #' @importFrom Seurat DotPlot FetchData
 #' @importFrom dplyr distinct bind_rows arrange desc top_n
@@ -133,6 +133,8 @@ Celltype_Verification <- function(
     }
   }
 
+  message(paste0(" "))
+  
   all_features <- unique(unlist(feature_list))
   if (length(all_features) == 0) stop("No valid features found for verification")