+++ R/Database_Cellmarker2.R
@@ -8,7 +8,7 @@
 #'     The dataset can be filtered based on species, tissue class, tissue type,
 #'     cancer type, and cell type to generate a list of marker genes for specific cell types.
 #'
-#' @family SlimR_Database
+#' @family Section_0_Database
 #'
 #' @source \url{http://117.50.127.228/CellMarker/}
 "Cellmarker2"