+++ R/Database_Cellmarker2_table.R
@@ -7,7 +7,7 @@
 #'
 #' @details This list is used to choose filters for creation of standardized marker list.
 #'
-#' @family SlimR_Database
+#' @family Section_0_Database
 #'
 #' @source \url{http://117.50.127.228/CellMarker/}
 "Cellmarker2_table"