+++ R/Database_Markers_list_scIBD.R
@@ -7,8 +7,10 @@
 #' @details This list is a table of 101 types of human intestine cell types markers
 #'     obtained from scIBD. The article doi source is "https://doi.org/10.1038/s43588-023-00464-9", 
 #'     and the reference literature is: Nie et al. (2023) <doi:10.1038/s43588-023-00464-9>.
+#'     Note: The 'Markers_list_scIBD' was generated using section 2.5.2 and the parameters 
+#'     'sort_by = "logFC"' and 'gene_filter = 20' were set.
 #'
-#' @family SlimR_Database
+#' @family Section_0_Database
 #'
-#' @source \doi{https://doi.org/10.1038/s43588-023-00464-9}
+#' @source \doi{10.1038/s43588-023-00464-9}
 "Markers_list_scIBD"