+++ R/Probability_calculate.R
@@ -19,7 +19,7 @@
 #' @returns Average expression of genes in the input "Seurat" object given
 #'     "cluster_col" and given "features".
 #'
-#' @family Use_in_packages
+#' @family Section_1_Functions_Use_in_Package
 #'
 #' @importFrom Seurat DefaultAssay DefaultAssay<- CellsByIdentities FetchData
 #' @importFrom stats sd weighted.mean