+++ R/onattach.R
@@ -1,15 +1,15 @@
 .onAttach <- function(libname, pkgname) {
   if (requireNamespace("crayon", quietly = TRUE)) {
-    message <- paste(
+    message <- paste0(
       "Please cite: Wang Z (2025). ",
-      crayon::italic("SlimR: Marker-Based Package for Single-Cell and Spatial-Transcriptomic Annotation."),
-      " R package version", crayon::bold("1.0.7."),
+      crayon::italic("SlimR: Machine Learning-Assisted, Marker-Based Tool for Single-Cell and Spatial Transcriptomics Annotation."),
+      " R package version", crayon::bold(" 1.0.8."),
       " Available at: https://github.com/Zhaoqing-wang/SlimR"
     )
   } else {
-    message <- paste(
-      "Please cite: Wang Z (2025). SlimR: Marker-Based Package for Single-Cell and Spatial-Transcriptomic Annotation.",
-      "R package version 1.0.7. Available at: https://github.com/Zhaoqing-wang/SlimR"
+    message <- paste0(
+      "Please cite: Wang Z (2025). SlimR: Machine Learning-Assisted, Marker-Based Tool for Single-Cell and Spatial Transcriptomics Annotation.",
+      "R package version 1.0.8. Available at: https://github.com/Zhaoqing-wang/SlimR"
     )
   }